Removing apply_to _original function

+++ tests/test_staging/test_validation.py
@@ -89,7 +89,7 @@ def test_validate_data_with_schema(mock_load_schema):
     validate_data_with_schema(dumy_data, "mock_schema.toml")
 
     # Check data types after validation
-    assert dumy_data["col1"].dtypes == int
+    assert dumy_data["col1"].dtypes == np.int64
     assert dumy_data["col2"].dtypes == pd.StringDtype()
     assert dumy_data["col3"].dtypes == float
     assert pd.api.types.is_datetime64_any_dtype(dumy_data["col4"].dtypes)
